requirements.txt
django
pytest
gunicorn
django-crispy-forms
crispy-bootstrap5
djangorestframework
requests
